# lib/api-zod/src/generated/types/childProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Fire TV discovery and parental safety demo API
 * OpenAPI spec version: 0.1.0
 */
import type { Thresholds } from './thresholds';

export interface ChildProfile {
  id: string;
  name: string;
  age: number;
  thresholds: Thresholds;
}