# lib/api-zod/src/generated/types/contentVerdict.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Fire TV discovery and parental safety demo API
 * OpenAPI spec version: 0.1.0
 */

export type ContentVerdict = typeof ContentVerdict[keyof typeof ContentVerdict];


export const ContentVerdict = {
  approved: 'approved',
  flagged: 'flagged',
  blocked: 'blocked',
} as const;