# lib/api-zod/src/generated/types/decisionRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Fire TV discovery and parental safety demo API
 * OpenAPI spec version: 0.1.0
 */
import type { DecisionRequestDecision } from './decisionRequestDecision';

export interface DecisionRequest {
  requestId: string;
  parentAccountId: string;
  decision: DecisionRequestDecision;
}