# lib/api-zod/src/generated/types/decisionResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Fire TV discovery and parental safety demo API
 * OpenAPI spec version: 0.1.0
 */
import type { DecisionResponseDecision } from './decisionResponseDecision';

export interface DecisionResponse {
  ok: boolean;
  decision: DecisionResponseDecision;
}