# lib/api-zod/src/generated/types/pairRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Fire TV discovery and parental safety demo API
 * OpenAPI spec version: 0.1.0
 */

export interface PairRequest {
  parentAccountId: string;
}