# lib/api-zod/src/generated/types/pairSession.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Fire TV discovery and parental safety demo API
 * OpenAPI spec version: 0.1.0
 */

export interface PairSession {
  code: string;
  parentAccountId: string;
  expiresAt: string;
}