# lib/api-zod/src/generated/types/parentAccount.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Fire TV discovery and parental safety demo API
 * OpenAPI spec version: 0.1.0
 */
import type { ChildProfile } from './childProfile';

export interface ParentAccount {
  id: string;
  name: string;
  children: ChildProfile[];
}