# lib/api-zod/src/generated/types/profilesResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Fire TV discovery and parental safety demo API
 * OpenAPI spec version: 0.1.0
 */
import type { ParentAccount } from './parentAccount';

export interface ProfilesResponse {
  parentAccount: ParentAccount;
}