# lib/api-zod/src/generated/types/searchResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Fire TV discovery and parental safety demo API
 * OpenAPI spec version: 0.1.0
 */
import type { Title } from './title';

export interface SearchResponse {
  source: string;
  results: Title[];
}