# lib/api-zod/src/generated/types/thresholdLevel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Fire TV discovery and parental safety demo API
 * OpenAPI spec version: 0.1.0
 */

export type ThresholdLevel = typeof ThresholdLevel[keyof typeof ThresholdLevel];


export const ThresholdLevel = {
  none: 'none',
  mild: 'mild',
  moderate: 'moderate',
  severe: 'severe',
} as const;