# lib/api-zod/src/generated/types/thresholds.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Fire TV discovery and parental safety demo API
 * OpenAPI spec version: 0.1.0
 */
import type { ThresholdLevel } from './thresholdLevel';

export interface Thresholds {
  violence: ThresholdLevel;
  language: ThresholdLevel;
  sexContent: ThresholdLevel;
  substances: ThresholdLevel;
  scaryContent: ThresholdLevel;
}