# lib/api-zod/src/generated/types/title.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Fire TV discovery and parental safety demo API
 * OpenAPI spec version: 0.1.0
 */

export interface Title {
  id: string;
  title: string;
  year: number;
  posterPath: string;
  overview: string;
  genres: string[];
  service: string;
}