# lib/api-zod/src/generated/types/titleSelectionResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Fire TV discovery and parental safety demo API
 * OpenAPI spec version: 0.1.0
 */
import type { ContentVerdict } from './contentVerdict';
import type { Title } from './title';

export interface TitleSelectionResponse {
  requestId: string;
  verdict: ContentVerdict;
  title: Title;
  reason: string;
  categoriesOfConcern: string[];
  service: string;
}